Include HAMA and ASS plug-ins
We want to include these plug-ins to our Plex service, but the
location where these should be placed does not exist until runtime.
We solve this by downloading them to an /extras folder on the side
and then symlink them to the correct location during startup.

This appears to work better than expected during initial testing.