Move Plex from a s6 service to a s6 CMD
In my opinion it is weird to run a single service under a supervisor
in Docker, but in order to not rewirte the entire image there is
sort of a recommended approach mentioned in the s6 docs regarding
cases like this.

If we only have a single service, that should bring down the container
if it fails, it is recommended that you provide this as a CMD argument
while still having /init as the ENTRYPOINT. Initial testing have me
feel like it behaves in a more sane way than before (where failing
startup scripts were ignored).